+++ internal/app/passwords_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/austin/go-unrarall/internal/rar"
+	"github.com/arodd/go-unrarall/internal/rar"
 	"github.com/nwaples/rardecode/v2"
 )