+++ internal/app/skip_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/austin/go-unrarall/internal/rar"
+	"github.com/arodd/go-unrarall/internal/rar"
 )
 
 func TestAlreadyExtractedFromListedFullPath(t *testing.T) {